requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djoser==2.2.2
pytest==8.3.4
pytest-django==4.9.0

